SnapFlingBehavior without Density
Remove density from SnapFlingBehavior. All implementations of SnapLayoutInfoProvider already have a way of accessing the density, the receiver scope could be removed which will lead to a less complex implementation of both SnapFlingBehavior and Snap

This means that the velocity threshold has to be provided in pixels/s and it's now a non-optional parameter to create SnapFlingBehavior outside of composition.

Test: N/A
Relnote: Remove density from SnapFlingBehavior. All implementations of SnapLayoutInfoProvider already have a way of accessing the density, the receiver scope could be removed which will lead to a less complex implementation of both SnapFlingBehavior and SnapLayoutInfoProviders.

Change-Id: I153c358968d8f4bfec027f0568f29c887aa06d53