[Unit test lib] Add `onAllNodes` method that returns a `GlanceNodeAssertionCollection` to perform assertions on collection of nodes. Supported assertions on collections include: `assertAll`, `assertAny`, `assertCountEquals`. Additionally, following operations can be performed on a collection: `filter` to filter the collection via a matcher, `get` to get node at an index in collection. Also adds a `onChildren` method that allows performing assertions on children of a node returned by `onNode`.
Bug: b/201779038
Test: "./gradlew :glance:glance-testing:testDebugUnitTest", "./gradlew :glance:glance-appwidget-testing:testDebugUnitTest"
Relnote: "Adds a `onAllNodes` method that can be called in a `runGlanceAppWidgetUnitTest` scope to find all nodes matching a matcher and performing assertions on them. Assertions supported on collections include: `assertAll`, `assertAny`, `assertCountEquals`. Additionally, following operations can be performed on a collection: `filter` to filter the collection via a matcher, `get` to get node at an index in collection. Also adds an `onChildren` method that can be called on result of calling `onNode` to get children of filtered node and perform assertions on them. "
Change-Id: If8d79b18c36d887e9552f8f7e54aefd6bdcb7e09