Add API for dropping all tables during destructive migrations.
The APIs are overloaded versions of the current fallbackToDestructiveMigration APIs with the older ones deprecated in order to have users migrate to the recommended behavior of dropping all tables, done so via the extra param in the overloaded version.

The API and behavior change addresses the issue of an application adding a new table in a version n+1 schema, then rolling back the deployment causing a downgrade but due to the table / entity not existing at version n Room would unknowingly keep the table (its untracked) thus causing a failure when the redeployment of the app at version n+1 is done again along with its migration.

Test: MigrationTest.dropAllTablesDuringDestructiveMigrations
Relnote: "Overload fallbackToDestructiveMigration APIs with a boolean parameter to indicate if all tables, including those not tracked by Room should be dropped during destructive migrations. Usually the value should be 'true' unless the database contains tables not defined via Room entities. The version of the APIs without the parameter are now deprecated and the default behavior of not destroying untracked tables is kept."
Change-Id: Ic8ef384aef7443430a1fc1e461aacaa18065126d