Updated Vector caching logic.
Updated cache to persist the fully parsed path
and group hierarchy that represents the root
GroupComponent instead of the VectorPainter instance.
On requests to load a vector asset a new VectorPainter
will be created that shares the same underlying datastructures
for the same vector asset. This prevents potential thrashing
of re-rendering the same vector asset if it is to be
drawn multiple times in the same scene at different resolutions.

Before: 342,810 ns          64 allocs
After: 13,980   ns           4 allocs

Bug: 295582473
Test: Updated VectorTest
Change-Id: I3cdf5aaafb7bf770ea1c2af5b7472cded2ee23dd