[M3] Introduce flag for Scaffold Subcompose Measure fix
aosp/2712234 introduced a fix for components relying on onSizeChanged to be called at a certain time. This change caused issues in some complex scenarios that need more investigation, so we are introducing a flag (`ScaffoldSubcomposeInMeasureFix`) to control this behavior.

We are aiming to remove it in Compose 1.6.0-beta01.

Test: scaffold_subcomposeInMeasureFix_enabled_measuresChildrenInMeasurement, scaffold_subcomposeInMeasureFix_disabled_measuresChildrenInPlacement
Relnote: Introduced a temporary flag to control whether Scaffold should measure its children during measurement or during placement. By default, this will measure in measurement. If you are facing issues with the new behavior, please file an issue.
Bug: 299282154
Change-Id: I0b354a595fa56d96c4d48cfae5b394c7203bd23c