[BottomSheet] Can override the maximum width of ModalBottomSheet and BottomSheetScaffold
Adding a new parameter, sheetMaxWidth, that can be used to specify the maximum width that the sheet would take up.
Dp.Unspecified can be passed in for sheetMaxWidth if a sheet that spans the entire width of the screen is desired.

Test: Adding a tests for ModalBottomSheet and BottomSheetScaffold that tests when Dp.Unspecified is passed in for sheetMaxWidth to ensure that the sheet spans the entire width, as well as different sheetMaxWidth values.
Bug: 266697696
Relnote: Adding a new sheetMaxWidth parameter that developers can set to specify a maximum width that the sheet will span. Dp.Unspecified can be passed in for the parameter if a sheet that spans the entire screen width is desired.
Change-Id: Ifb7c9ee4d0066e86787e8fcbf0d156b9f92e5cfb