Implement capability exchange onCallAdded.
This CL adds the logic that triggers capability exchange when InCallService#onCallAdded is invoked. It leverages InCallServiceCompat#resolveCallExtensionsType to determine if capability exchange is supported and if it is. This CL also removes InCallServiceCompat's dependence on GlobalScope and instead creates its own coroutinescope when it is created which is destroyed when that ICS is destroyed.

Fixes: 299188162
Test: CallCompatTest + E2ECallExtensionsExtrasTest
Change-Id: I9587b232582011ae818dae25ddd7ea341403bd08