Reduce delay between a layout becoming visible and its pipeline nodes being initialized.
Attach a new layout to the parent view before pipeline init, so that we don't have to post the init call in the view.
Force a measure pass to happen if there's any SlideIn animations in the new nodes.

Relnote: Reduce delay between a layout becoming visible and its pipeline nodes being initialized.

BUG: 249019163
BUG: 285570489
Change-Id: I38d4e8cd73037e5099f1b88f595b659cbc8436a5