Do not leak effect Job in SessionWorker
aosp/2729893 introduced an independent Job for composition effects so
that we can catch errors. This Job continues to run after the worker
ends because cancellation does not propagate automatically for
independent Jobs.

This change cancels the effect job on completion of the TimerScope, so
that it won't leak in any case.

Bug: 295250045
Test: SessionWorkerTest.sessionWorkerDoesNotLeakEffectJobOnCancellation,
 SessionWorkerTest.sessionWorkerDoesNotLeakEffectJobOnClose
Change-Id: Icc29f1925ba16e2c627be76a693009d33e763a9c