Add Fade-in animation for PositionIndicator. Also add new animation flags into new PositionIndicator api.
Fade-in animation for PositionIndicator was added back.
3 separate flags were added to control different animations of PositionIndicator.
- showFadeInAnimation - controls fade-in animation.
- showFadeOutAnimation - controls fade-out animation.
- showPositionAnimation - controls position change animation.
Old api without this flag was deprecated.

For performance reasons and for UX consistency, when PositionIndicator is used with scrollable list, we recommend to switch off showFadeInAnimation and showPositionAnimation flags.
If PositionIndicator is used as a standalone indicator, for example for volume change, then we recommend to have all 3 animations turned on.

Bug: 301580348
Test: PositionIndicatorBenchmark.kt
Relnote: "3 separate flags were added to control different animations of PositionIndicator. showFadeInAnimation - controls fade-in animation. showFadeOutAnimation - controls fade-out animation. showPositionAnimation - controls position change animation. Old api without this flag was deprecated. For performance reasons and for UX consistency, when PositionIndicator is used with scrollable list, we recommend to switch off showFadeInAnimation and showPositionAnimation flags. If PositionIndicator is used as a standalone indicator, for example for volume change, then we recommend to have all 3 animations turned on."
Change-Id: Ia2d63d8ea92ca87b09b11d79c7968f0f72460bab