Add exit condition for CaptureSessionStuckQuirk
Treat onCaptureCompleted and onCaptureFailed as a part of exit
condition.

onCaptureCompleted is usually called after onCaptureStarted
onCaptureCompleted and onCaptureFailed do not usually trigger the task to be complete, because onCaptureStarted may already have triggered it.
onCaptureCompleted and onCaptureFailed are used to ensure that the ListenableFuture is completed when the capture is aborted. When abortCaptures() is called, onCaptureCompleted or onCaptureFailed will be triggered, but onCaptureStarted may not be triggered.

Test: ./gradlew bOS && Camera2CameraImplTest
Change-Id: Idd33e5c2f1a9f224156f5694888dbaa3c9090777