Fix issue with disappearing PositionIndicator (ScrollBar) in some cases
The problem happens when we add top and bottom paddings to the list and total height of items is shorter than the height of the viewport.
Solution is to use canScrollForwards and canScrollBackwards methods instead of relying on start/end offsets.

Bug: 301904803
Test: PositionIndicatorTest.kt
Relnote: "Fix issue when PositionIndicator was disappearing in some cases"
Change-Id: I2091a88014bbaccf5ff6171d9bf2c50c96d4ae36