Extend FontStyle#size to accept a list of sizes
This updates the existing proto field to accept a list of sizes that
will be used to automatically scale the text to best fit.

The API to set multiple sizes is experimental.

Renderer implementation will be done in a follow up. Currently, it's
only changed to address the change in proto field.

Bug: 302531969
Test: Unit tests for wrappers
Relnote: "We have added an experimental API to automatically scale the text size based on the space it has inside of the parent."
Change-Id: Ibbe632447dcf28a9948a0877f193b4912acb23d1