Deprecate SwipeToDismissBoxState, SwipeToDismissValue and edgeSwipeToDismiss
These can now be deprecated as all Swipe-to-Dismiss functionality is provided by wear.compose.foundation. In particular, the Wear.Compose.Material.SwipeToDismissBoxState had a foundationState property which was raised as a concern in the api-review - this only existed for backwards compatibility and deprecating these classes makes that clearer.

Bug: 292537483
Test: Updated tests to replace deprecations
Relnote: "We have deprecated the Material SwipeToDismissBoxState, SwipeToDismissValue and edgeSwipeToDismiss following the migration of Swipe-to-Dismiss functionality to wear.compose.foundation. Please replace with the wear.compose.foundation equivalents."
Change-Id: Iee8c97aeb50a9109d35fc70d41c57fb755dab699