Fix crash when creating LifecycleCamera with a destroyed lifecycle
The LifecycleStatusChangeStressTest may quickly open and close
the Activity. This can cause the Activity to be in a destroyed state
when the LifecycleCamera is created. This can lead to a crash.

To prevent this crash, the lifecycle status of the Activity should
be checked before creating the LifecycleCamera. If the Activity is
in a destroyed state, the LifecycleCamera should not be created.

Bug: 303340037
Test: LifecycleStatusChangeStressTest
Change-Id: Id3b2ede0bbd0f69f0dfd1eeb89550aef0d7def98