Make LoremIpsum PreviewParameterProvider open
Since PreviewParameterProviders are passed as a class object to
PreviewParameter annotations, it's not possible for users to set
different constructor parameter. The best way to achieve that is
creating a subclass of this provider that sets the word count, however
this can only be done if the class is open.

Fixes: 266918816
Fixes: 300116360
Test: PreviewParameterTest#checkLoremIpsumProviderSubclass
Relnote: Made LoremIpsum PreviewParameterProvider an open class
Change-Id: I41bf5f4e880867aea3603b3c4ea9a37c456e428d