Cursor, cursor handle and selection handles will now disappear when window loses focus
When textfield is in focus but window loses its focus, the textfield will preserve its focus but it will stop animating the cursor and will dismiss the cursor/selection handles.
If there's a selection, it will continue to be drawn (as a background) but handles/toolbar will disappear.

This also fixes a bug where previously toolbar never appeared back after window regained focus because we didn't restart the input session.

There're a lot of changes in the tests but most of them (apart from those that test the new functionality) simply updated to use the `rule.setTextFieldTestContent` which propagates the WindowInfo with the isWindowFocused==true.

Bug: 170977431
Test: multiple new tests + a demo

Change-Id: Ia09393e503061c39d6b909981caf723b500e3588