Selection correctly shrinks in touch selection when a word spans multiple lines
Selection was stuck in a loop where it thought that every selection change crossed lines, resulting in word based adjustment being used. The selection result was at the line end, but the line end counted as the beginning of the next line, and thus when we queried the line number for the offset, it would be the next line and not the line we are currently hovering. This change instead compares the line of the raw previous offset (before selection adjustment), instead of the adjusted previous offset, so that the line change is correctly determined.

Test: ./gradlew :com:f:f:cAT
Fixes: b/265636253
Change-Id: I786d74aa3747507474c9abef531f1102dc9f4fc8