Explictly layout the child view in LocalAdapter#notifyResized
This change matches the resize behavior that exists with the remote
adapter. SandboxedSdkView delegates the resizing to the adapter factory
in the resize flow, so this needs to be handled by the local adapter
in the backwards compat case.

Adds more end to end tests for validating app and provider-initiated
resizes.

Test: Manual
Test: IntegrationTests
Bug: 303387300
Change-Id: Iff25e24d939b924affca6dfff27d91b49e3f05f9