Fix ViewModel not cleared after pop
When an entry is popped followed by a configuration change, the list of transitioning entries is lost and the outgoing entry is not marked complete. As a result, the ViewModel is also not cleared.

Now we ensure to mark outgoing entries as complete when a NavHost goes through config change (when it is disposed).

Test: ./gradlew navigation:navigation-compose:cC
Bug: 298164648
Relnote: "When an entry is popped followed by a configuration change, the entry's ViewModel will now be cleared properly if saveState is false."
Change-Id: Idf242239ceeb731c3375ef507a2cc854d85743c7