Hide PositionIndicator by default
A behavioral change for PositionIndicator was introduced in previous CLs (aosp/2747773, aosp/2758367, aosp/2775142, aosp/2770702), which was showing PositionIndicator by default before the list is scrolled. That behavior was reverted. As a consequence, AutoHide isn't showing PositionIndicator by default, it only hides it.

Bug: 304743797
Test: PositionIndicatorTest.kt
Relnote: "Hide PositionIndicator by default. A behavioral change for PositionIndicator was introduced in previous CLs, which was showing PositionIndicator by default before the list is scrolled. That behavior was reverted."
Change-Id: I41843ef10d4ebad839616c8005070f1989b2e940