Fix NPE for private display
This is a fix for the regression issue caused by
Ie65445d36be6b771543a30263c66252adf8bed84

A Display object is not always accessible from uiautomator because the
display can be private.
Even if the display is not accessible, uiautomator shouldn’t throw NPE.
Apps are not allowed to place windows upon private displays, but it
doesn’t mean we can’t interact with the content on it and that’s why
accessibility tree contains the objects on it.

Bug: 303752326
Test: atest GtsIncidentManagerTestCases
Change-Id: I816f48ff89e75eea472081feed1ca72c14cf588f