Allow NavHost to display entries on first composition
Currently on the first composition the NavHost does not have an entry to
display and it has to wait for the next composition for the state to be
updated.

By rearranging the code, we can get a non-empty back stack on the first
composition.

RelNote: "NavHost will not render the start destination on the first
composition pass."
Test: all tests including preview pass
Bug: 304852206

Change-Id: I439a7fe894907b1ebee251134113acb08973083c