Yet another attempt at fixing test flakiness
This time we take a different approach. The main issue we seem to
have is that when we take a screenshot, our SurfaceView's surface
isn't ready yet, indicating the window manager hasn't had time
to do its thing, punch the hole, position the sub-window, etc.

Since we don't have signals from WindowManager, this change adds
its own window, waits for its content to be drawn and *then* runs
the rest of the test. Since we add our window *after* the
SurfaceView is laid out, this should in theory guarantee that
the SurfaceView's Surface window is on screen.

Fixes: 302165563
Test: AndroidExternalSurfaceTest
Relnote: N/A
Change-Id: Idbf771b8d6fc6b3123a71e38510e1dbe8ef0063f