Fix moveable content sometimes receiving the wrong composition locals
In rare cases the composition locals scope would be incorrectly
calculatedfor moveable content. This cases are hard to get into as it
requires the slot table groups of the source and destination to both
provide a composition locals scope at the same group address and have
the destination record and update to that group in the same composition
that is also composing the movable content.

The fix is to not use the provider update map of the destination table
when recomposing moveable content.

Also include an update to MutableIntIntMap for node counts to fix a
new lint error.

Fixes: 304279025
Test: Ad hoc. Any test case for this would quickly become out of date.
Change-Id: I99bccb7c7ef6b60ea07512c1c4d25399f9dec58f