Coerce how replace change is tracked by EditingBuffer
EditingBuffer always reports replace calls as is to ChangeTracker. However, some replace calls are essentially does not replace the applied region but insert or delete. For example composition based keyboards always call replace for the composing word even though user is typing characters one by one. This makes it very difficult to track changes in UndoManager.

Coercing the ends of the replace call until an actual change is found helps us to pinpoint the exact change happening in the buffer.

Test: :compose:foundation:foundation:cAT
Change-Id: Ia293c3ffd3aa607ec5c38249de5678263f5b6528