Increase accurate tool jank for motion predictor
We have received a report that the prediction can be very
conservative at high speeds; after analising the jank values
at high speed, it seems that the threshold that was picked (0.2)
is too low.

This CL increases the value to 0.7 since it seems to be good for
high speed straight lines. A better way to computer perceived
errors will be implemented later.

Bug: 232941452
Test: Built, tested on sample app
Change-Id: Ic9509f7de25b27c23c62b8962c09c32d6a39d1df