[Slider] Updating FloatRange with API council feedback
Updates the FloatRange class in Slider.kt with the API Council feedback. Since this class is now targeted for RangeSlider, removing the usage of FloatRange in SnapFlingBehavior.kt.

Bug: 295321399
Test: N/A
Relnote: Updating Slider.kt with the API council feedback for FloatRange. Removing the use of this class in SnapFlingBehavior.kt since it's specific to RangeSlider now.
Change-Id: I025cb6717a8f59e4ce70a8017fb76594383489ce