Reapply "Handle AppSearch polymorphic deserialization correctly" (v2)
This reapplies the change aosp/2708413, but in addition, we never automatically build `AppSearchDocumentClassMap`. Instead, we provide a new overloaded `GenericDocument.toDocumentClass()` method that accepts a map. In this way, users can build the map at an appropriate time or even pass a custom map on their own.

Patchset 1 is uploaded as a pure reapply of aosp/2708413, which helps to review what's changed after aosp/2708413.

Bug: 290389974
Test: Presubmit
Relnote: Provide a new overloaded `GenericDocument.toDocumentClass()` method that accepts a map
Change-Id: I7619226406b665e39bea6e68dd94546568bf051a