Use larger icon size for Window's IconImage (#890)
This avoids blurry icons on window switchers (Alt+Tab) on Windows and
Linux.

## Proposed Changes

- change icon image size for windows to a larger base image 192px
instead of 32px, otherwise we end up with blurry icons (see
https://github.com/JetBrains/compose-multiplatform/issues/1838)

## Testing

Test: I have not tested this change directly, but am using this size on
other apps to set the size and this seems to work well on the systems I
tried on.

## Issues Fixed

Fixes: https://github.com/JetBrains/compose-multiplatform/issues/1838

## Google CLA
You need to sign the Google Contributor’s License Agreement at
https://cla.developers.google.com/.
This is needed since we synchronise most of the code with Google’s AOSP
repository. Signing this agreement allows us to synchronise code from
your Pull Requests as well.