Fix `ShaderBrush.createShader` state read not redrawing Text
When given brush is a custom implementation of `ShaderBrush` which reads a snapshot state in its `createShader` function, eager caching that only invalidates on size changes fails to redraw text to reflect the latest shader. This fix reverts the `setBrush` logic in `AndroidTextPaint` to a previous implementation while adding `derivedStateOf` wrapper to make sure that state reads re-sets the internal shader.

Fix: b/310955096
Test: :compose:found:found:cAT
Change-Id: I1c7ecaec36aaa30816c08cf5df01dc3b3d508af7