Remove temporary SurfaceView logic on T-
There is no need to attach a temporary SurfaceView to a
SandboxedSdkView on T-, as the remote adapter will not be
used in this case. Instead, we can directly set the window input
token to a new Binder object on T- devices.

This change also means that we do not need to create a SandboxedSdkView
on older SDK versions on the UI thread, as we no longer need to create
a SurfaceView which requires one on these SDKs.

Test: IntegrationTests, SandboxedSdkViewTest, run on U and N
Bug: 308751186
Bug: 311151730
Change-Id: I31027717b3ff4f39bd73e222ffbde9018d8409f0