ScatterMap.asMap() iterator now return new instances
Clients of ScatterMap.asMap() may rely on the map iterator returning new
instances instead of the same wrapper Entry instance. This is the case for
instance of the toList() utility function in Kotlin's stdlib. As a result,
calling ScatterMap.asMap().toList() would always yield a list containing
the *same* Pair in every slot.

This change forces the map iterators to return new entry instances instead.
It's wasteful from a memory/GC's perspective but it matches existing
expectations and code.

Fixes: 310365754
Test: ScatterMapTest
RelNote: N/A
Change-Id: I2893260c67fcbd3388c74043786c0329570ca5db