Optimize cubic roots finding and evaluation a little bit
With the switch to exact computations, cubic easing showed a performance regression in
benchmarks. On a Pixel 2, the animation2D benchmark went from 18,500 ns to 24,100 ns.
After these optimizations, the benchmark runs in 22,020 ns.

The rest of the difference is because of the use of sqrt() and 2x cbrt() in the root
finding function. Without those, the benchmark runs in 11,500 ns. Unfortunately we need
them and considering they allow us to have an allocation-free CubicEasing implementation
that's independent of the time duration and provides almost exact results, the extra
cost is worth it (and similar to/cheaper than the cost of interpolating 2 colors through
OkLab for instance).

Fixes: 311552672
Test: EasingTest (instrumented and host), PathEasingTest
RelNote: N/A
Change-Id: I809dff4c67606ac6c7d1d89df8ef21946f6dc5dd