Fix flaky tests in JankStatsTest
Some of the tests in JankStatsTest run afoul of race
conditions between the UI/test thread and a separate
thread that is inserting frame metrics data when frames are
processed. This makes it tricky to test for deterministic results,
leading to flaky tests, especially when run on servers with
much different CPU characteristics than local hosts.

The fixes here are varied. In some cases, it was possible to
make things more detemrinistic by, for example, ensuring that
states are only inserted or removed after ensuring that older frames
have already been processed. In other cases, the tests were simplified
to avoid essentially untestable situations (or at least situations
that cannot be tested down to the per-frame level, which is what
they were doing previously).

Also, updated library version in preparation for releasing first beta.

Bug: 297129762
Bug: 296918479
Bug: 238890043
Test: Ran JankStatsTest locally and on test dashboards

Change-Id: Iba11a91494f19b76e06266fb0569f60a97f48df1