Split VirtualCamera into VirtualCameraAdapter  and VirtualCamera
Previously, there is a 1:1 relationship between StreamSharing and VirtualCamera based on the assumption that the StreamSharing does not perform rotations. Now that we want to perform the rotations, this assumption is no longer true, because each child's virtual camera may have different rotations.

In this CL, we break up VirtualCamera into VirtualCameraAdapter and VirtualCamera. VirtualCameraAdapter works like the old VirtualCamera. It manages the states of the UseCases which is similar to CameraUseCaseAdapter, hence the name "Adapter". VirtualCamera is a CameraInternal implementation created specificaly for a child. It allows customizing the rotation degrees so that when the child call getSensroRotationDegrees(int), it would return the correct value.

- In VirtualCameraInfo, return a rotation degree based on the rotation performed in the sharing node.
- In Preview, removed the special casing for shared stream.
- Fixed the landscape layout of EffectsFragment.

Bug: 310260193
Test: manual test and ./gradlew bOS
Change-Id: Ic8c4481fe120603295283597f34c00bbac0f4bb3